merge preprocessCall into the existing handlePreCall stuff
This is more complex than it looks because that translation wants
to happen at the top level, and because the translation for
Distinct is only right if it's in a bitmap context -- in other
contexts, we don't want to translate it into the naive Precomputed
positive-values-only. And currently that means actually executing
it live, but it *should* mean still precomputing it, but using
a SignedRow.

So we in practice end up with two special cases, both of which
are fundamentally *about* Distinct:
* Distinct calls don't want the precall computation
* Count() calls don't want precall on their immediate children
* Everything else *does* want the precall computation

Before this change, we didn't actually need the precall on the
top level things, but now we want it because the top level thing
could be an All(limit=1) which we want to translate.

This needs some reworking. To really fix this, though, we have
to fully switch over to a fancier setup with an analyze phase
and the map/reduce/finalize functions. And looking at GroupBy,
in particular, it's now clear that those functions can't
always be static -- they may need to be generated as closures
that reflect additional precomputation work.