postgres or bolt, postgres in CI, fix invalidtx errors, require
Changes the controller setup to allow postgres or bolt mode. This
isn't tested yet.

Tries getting postgres running in CI so tests can run in CI

InvalidTransaction error was hardcoded to boltdb... made that flexible

Changed testify/assert to require in a lot of places so tests don't
continue executing when things are clearly broken